requirements.txt
Flask
flask-cors
torch
numpy
Pillow
gunicorn